refactor(types): progress
