+++ src/lib/builders/progress/types.ts
@@ -1,0 +1,8 @@
+import type { createProgress } from './create';
+
+export type CreateProgressArgs = {
+	value?: number;
+	max?: number;
+};
+
+export type CreateProgressReturn = ReturnType<typeof createProgress>;
